requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
agno
jinja2
google-genai
httpx
gunicorn
tavily-python
pymupdf4llm
motor
pymongo
faiss-cpu
pypdf
groq
numpy
faiss-cpu
pypdf
